task_2_nlp/requirements.txt
datasets==4.2.0
kagglehub==0.3.13
matplotlib==3.10.7
pillow==11.3.0
ruff==0.7.4
scikit-learn==1.7.2
seqeval==1.2.2
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.0
